FEAT: Create task panel header
